maintain filter after sort

+++ app/controllers/movies_controller.rb
@@ -14,10 +14,10 @@ class MoviesController < ApplicationController
     @all_ratings = Movie.all_ratings
     @movies = Movie.all
     
-    if params[:ratings] == nil
+    if params[:ratings] == nil and params[:sort] == nil
       @selected = Movie.all_ratings
       session.delete(:ratings)
-      params[:rating] = ['G' => 1, 'PG' => 1, 'PG-13' => 1, 'R' => 1, 'NC-17' => 1]
+      params[:ratings] = {'G' => 1, 'PG' => 1, 'PG-13' => 1, 'R' => 1, 'NC-17' => 1}
     end
     
     if params[:ratings] != nil
